requirements.txt
# PROJECT LUMEN - Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator>=2.0.0

# Document Processing
pdfminer.six==20221105
pytesseract==0.3.10
Pillow==10.1.0

# RAG & Embeddings
sentence-transformers>=2.7.0
faiss-cpu==1.9.0
rank-bm25==0.2.2

# NLP & ML
transformers==4.35.2
torch==2.5.0
scikit-learn==1.3.2
numpy>=1.24.3

# LLM Integration
openai==1.3.7
google-generativeai>=0.3.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Image Forensics (Phase B)
opencv-python>=4.8.0
scikit-image>=0.22.0

# Optional but recommended
# For better OCR results (requires Tesseract OCR installation)
# tesseract-ocr (system package)

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
